Updated 1/10/2015
+globalvalues.py
Moved some code from snowball_adv.py so that it can be called in other
locations without causing error in python (endless import loops)

- snowball_adv.py
Moved some code to globalvalues.py. Check changes for that for more
details

+ syscombat.py
Changed some values in order to accomodate for the changes.
Some extra code that will help with the generation of enemychoice()